backend/requirement.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=3.14.0
requests>=2.26.0
supabase>=0.1.0
psycopg2>=2.9.3  # PostgreSQL adapter for Python
python-dotenv>=0.19.2  # For loading environment variables from a .env file